lib update and stderr logging
